pyproject.toml
[tool.poetry]
name = "cryopicls"
version = "0.1.0"
description = "Cryo-EM Particle Image Clustering in the Latent Space"
license = "MIT"
authors = ["Kotaro Tanaka <kttn8769@gmail.com>"]
readme = "README.md"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python :: 3",
]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
matplotlib = "^3.3.4"
numpy = "^1.20.1"
pandas = "^1.2.2"
plotly = "^4.14.3"
dash = "^1.19.0"
dash-bootstrap-components = "^0.11.3"
scikit-learn = "^0.24.1"
umap-learn = "^0.5.1"
PyYAML = "^5.4.1"
pyclustering = "^0.10.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.8.4"
yapf = "^0.30.0"

[tool.poetry.scripts]
cryopicls_clustering = "cryopicls.cryopicls_clustering:main"
cryopicls_projector = "cryopicls.cryopicls_projector:main"
cryopicls_visualizer = "cryopicls.cryopicls_visualizer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
